Move shared types into another namespace
