+++ source/ProcessManager.Orchestrations/ProcessManager.Orchestrations.csproj
@@ -13,7 +13,7 @@
   <ItemGroup>
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
     <PackageReference Include="Energinet.DataHub.Core.Messaging" Version="6.1.2" />
-      <PackageReference Include="Energinet.DataHub.ElectricityMarket.Integration" Version="3.0.0"/>
+      <PackageReference Include="Energinet.DataHub.ElectricityMarket.Integration" Version="3.0.0" />
     <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.ServiceBus" Version="5.22.0" />
     <PackageReference Include="Microsoft.Azure.Functions.Worker.Sdk" Version="2.0.0" />
     <PackageReference Include="Energinet.DataHub.Core.App.FunctionApp" Version="14.0.2" />
