+++ source/ProcessManager.Orchestrations/Processes/Shared/BRS_026_028/SearchTrigger_Brs_026_028.cs
@@ -12,19 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_026;
+using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.Shared.BRS_026_028;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Azure.Functions.Worker;
 using FromBodyAttribute = Microsoft.Azure.Functions.Worker.Http.FromBodyAttribute;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_026;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.Shared.BRS_026_028;
 
-// TODO:
-// Should be moved to another namespace because this is shared between BRS 026 + 028.
-// We have talked about combining these BRS's into ne top-folder similar to BRS 023 + 027,
-// and then use subfolders to split them per orchestration OR perhaps even use the same orchestration
-// because their logic is very similar.
 internal class SearchTrigger_Brs_026_028(
     SearchActorRequestHandler handler)
 {
